refactor: Move SubjectData and ViconFrame to types.hpp
